requirements.txt
ultralytics==8.2.84
opencv-python==4.10.0.84
Pillow==10.4.0
numpy>=1.24.0
scipy==1.13.1
